refactor: rename MetaNames to MetaNamesSdk
